Update kits build script
KiCad 5.1.9 embedded python was not built with zlib library so we need
to split the script to be able to compress the gerber folder artifact.
